refactor(backend): move dashboard business logic from repository to service
Addresses PR #119 review comments r3107510869 and r3107512731 from @Jorenve:
repository now holds only data-access (queries, counts, range lookups,
member reads); the service owns all derived math — delta/%-change,
month-bounds, locale-status classification, recent-item shaping, and
pagination meta. Public service contract and HTTP response shape are
unchanged.
